more test and add bsconsole.

+++ bsrouter/bsrouter-install.sh
@@ -1,14 +1,20 @@
 #!/bin/bash
 case "$1" in
   -i)
-    useradd bsrouter
-    mkdir -p /home/bsrouter
-    chown -R bsrouter:bsrouter /home/bsrouter
+    if [ ! -d /home/bsrouter ];then
+      useradd bsrouter
+      mkdir -p /home/bsrouter
+      chown -R bsrouter:bsrouter /home/bsrouter
+    fi
     cp -f bsrouter /usr/local/bin/bsrouter
     cp -f bsconsole /usr/local/bin/bsconsole
-    cp -f bsrouter.service /etc/systemd/system/
+    if [ ! -f /etc/systemd/system/bsrouter.service ];then
+      cp -f bsrouter.service /etc/systemd/system/
+    fi
     mkdir -p /etc/bsrouter
-    cp -f bsrouter.json /etc/bsrouter
+    if [ ! -f /etc/bsrouter/bsrouter.json ];then
+      cp -f bsrouter.json /etc/bsrouter
+    fi
     systemctl enable bsrouter.service
     ;;
   *)
